change JPA repo parameter of blood type from Integer to Long

+++ backend_blood_donation_system/backend_blood_donation_system/src/main/java/com/example/backend_blood_donation_system/repository/BloodTypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface BloodTypeRepository extends JpaRepository<BloodType,Integer> {
+public interface BloodTypeRepository extends JpaRepository<BloodType,Long> {
 
     Optional<BloodType> findByType(String type);
